Moved WorkflowException hirerarchy to client package. Added reconnect on execute from external
