+++ src/main/java/com/uber/cadence/client/WorkflowException.java
@@ -14,9 +14,10 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.workflow;
+package com.uber.cadence.client;
 
 import com.uber.cadence.WorkflowExecution;
+import com.uber.cadence.workflow.ChildWorkflowException;
 
 /**
  * Base exception for all workflow failures returned by an external client.
