+++ src/main/java/com/uber/cadence/internal/DynamicWorkflowClientExternalImpl.java
@@ -19,6 +19,7 @@ package com.uber.cadence.internal;
 import com.uber.cadence.ChildPolicy;
 import com.uber.cadence.WorkflowExecution;
 import com.uber.cadence.WorkflowType;
+import com.uber.cadence.client.WorkflowExecutionAlreadyStartedException;
 import com.uber.cadence.client.WorkflowOptions;
 import com.uber.cadence.converter.DataConverter;
 import com.uber.cadence.converter.JsonDataConverter;
